tests/requirements.txt
# 核心科学计算库
torch>=1.8.0
torchvision>=0.9.0
torchaudio>=0.8.0

# 数据处理和分析
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 配置和日志
PyYAML>=5.4.0

# 系统监控
psutil>=5.8.0

# 异步网络处理
aiohttp>=3.7.0
asyncio-mqtt>=0.11.0

# 可视化（可选）
matplotlib>=3.4.0
seaborn>=0.11.0

# 其他工具
requests>=2.25.0
docker>=5.0.0